requirements.txt
# Modules requis pour NOVA - Version harmonisee MCP-compatible
fastapi>=0.111.0
uvicorn[standard]>=0.31.1
httpx>=0.26,<1
anthropic>=0.8.0
openai>=1.6.0
simple-salesforce==1.12.4
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.5.3,<3
aiofiles==23.2.1
python-multipart>=0.0.9,<0.1.0
structlog==23.2.0
prometheus-client==0.16.0
tenacity==8.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
thefuzz[speedup]==0.22.1
alembic
starlette>=0.37.2
anyio>=4.5.0
mcp>=1.13.0
fastmcp>=0.1.2

# Parsers pour tarifs fournisseurs
pymupdf>=1.23.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pytesseract>=0.3.10
pillow>=10.0.0

# SQL Server connection pour fonctions SAP personnalisées
pyodbc>=5.0.0
